chore: refactor TD004 into multiple different plugins
