+++ test/fixtures/resources/TD011/pass/foo.xml
@@ -1,0 +1,13 @@
+<TargetEndpoint name="default">
+  <HTTPTargetConnection>
+    <SSLInfo>
+      <Enabled>true</Enabled>
+      <IgnoreValidationErrors>false</IgnoreValidationErrors>
+      <TrustStore>truststore1</TrustStore>
+    </SSLInfo>
+    <LoadBalancer>
+      <Server name="server1"/>
+      <Server name="server2"/>
+    </LoadBalancer>
+  </HTTPTargetConnection>
+</TargetEndpoint>
