+++ test/fixtures/resources/TD011/pass/t2-ignore-not-present.xml
@@ -1,0 +1,9 @@
+<TargetEndpoint name="default">
+  <HTTPTargetConnection>
+    <SSLInfo>
+      <Enabled>true</Enabled>
+      <TrustStore>truststore1</TrustStore>
+    </SSLInfo>
+    <URL>https://foo.com/apis/{api_name}/maskconfigs</URL>
+  </HTTPTargetConnection>
+</TargetEndpoint>
