+++ test/fixtures/resources/TD012/fail/t2-missing-SSLInfo.xml
@@ -1,0 +1,5 @@
+<TargetEndpoint name="default">
+  <HTTPTargetConnection>
+    <URL>https://foo.com/apis/{api_name}/maskconfigs</URL>
+  </HTTPTargetConnection>
+</TargetEndpoint>
