+++ test/fixtures/resources/TD012/pass/t2-no-SSLInfo-but-insecure-url.xml
@@ -1,0 +1,5 @@
+<TargetEndpoint name="default">
+  <HTTPTargetConnection>
+    <URL>http://foo.com/apis/{api_name}/maskconfigs</URL>
+  </HTTPTargetConnection>
+</TargetEndpoint>
