+++ test/fixtures/resources/TD013/fail/t2-missing-keyalias.xml
@@ -1,0 +1,14 @@
+<TargetEndpoint name="default">
+  <HTTPTargetConnection>
+    <SSLInfo>
+      <Enabled>true</Enabled>
+      <ClientAuthEnabled>true</ClientAuthEnabled>
+      <TrustStore>truststore1</TrustStore>
+      <KeyStore>keystore1</KeyStore>
+      <!--
+          <KeyAlias>key1</KeyAlias>
+      -->
+    </SSLInfo>
+    <URL>https://foo.com/apis/{api_name}/maskconfigs</URL>
+  </HTTPTargetConnection>
+</TargetEndpoint>
