+++ test/fixtures/resources/TD013/pass/t1-all-good.xml
@@ -1,0 +1,12 @@
+<TargetEndpoint name="default">
+  <HTTPTargetConnection>
+    <SSLInfo>
+      <Enabled>true</Enabled>
+      <ClientAuthEnabled>true</ClientAuthEnabled>
+      <TrustStore>truststore1</TrustStore>
+      <KeyStore>keystore1</KeyStore>
+      <KeyAlias>key1</KeyAlias>
+    </SSLInfo>
+    <URL>https://foo.com/apis/{api_name}/maskconfigs</URL>
+  </HTTPTargetConnection>
+</TargetEndpoint>
